backend/requirements.core.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
httpx>=0.25.1
pydantic>=2.0.0
python-dotenv>=1.0.0
optuna>=3.4.0
ccxt>=4.1.0
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
pyjwt>=2.8.0
email-validator>=2.0.0
cryptography>=41.0.0
alembic>=1.12.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pyotp>=2.9.0
qrcode[pil]>=7.4.0
psutil>=5.9.0
joblib>=1.3.2
tqdm>=4.66.1
